requirements.txt
Flask==2.3.3
requests>=2.31.0
openai>=1.3.7
markdown2>=2.4.0
azure-functions